# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_igeargeek_sales-system-api_internal_models.ActiveArchivedStatus:
    enum:
    - active
    - archived
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusArchived
...
  github_com_igeargeek_sales-system-api_internal_models.Contact:
    properties:
      company_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deleted_by:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      role_title:
        type: string
      status:
        $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.ActiveArchivedStatus'
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handlers.contactForm:
    properties:
      company_id:
        type: integer
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      role_title:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: |-
    CRM/sales-pipeline API. This spec currently covers the
    Prospect stage, Pipeline stage, and Dashboard summary routes
    as a scaffold — the annotation pattern on those handlers
    (see internal/handlers/prospect_stages.go, pipeline_stages.go,
    dashboard.go) extends to the rest of internal/routes/routes.go
    as they're annotated. Run `go run github.com/swaggo/swag/cmd/swag@v1.16.6
    init -g cmd/api/main.go -o docs --parseDependency --parseInternal` from the
    repo root to regenerate docs/ after editing annotations (the two flags are
    required to resolve gorm.DeletedAt across package boundaries).
  title: Sales System API
  version: "1.0"
paths:
...
  /contacts:
    get:
      description: 'Paginated, filterable Contact list. Filters: company_id, status,
        tag, search (name/email).'
      parameters:
      - description: Filter by Company ID
        in: query
        name: company_id
        type: integer
      - description: active or archived
        in: query
        name: status
        type: string
      - description: Filter by Contact tag
        in: query
        name: tag
        type: string
      - description: Search by name or email
        in: query
        name: search
        type: string
      - description: Sort field, prefix - for descending (created_at, name, email,
          company_name)
        in: query
        name: sort
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated contact list (data, page, per_page, total)
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List contacts
      tags:
      - contacts
    post:
      consumes:
      - application/json
      description: Creates a Contact. company_id and name are required; role_title
        must match an active configured job title (see /admin/job-titles).
      parameters:
      - description: Contact fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.contactForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.Contact'
        "400":
          description: Invalid body, missing fields, or invalid role_title
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a contact
      tags:
      - contacts
  /contacts/{id}:
    delete:
      description: Soft-delete (AuditedModel) — recoverable via Restore/Trash below.
        Never a hard delete, since Deal/Activity/Task records reference contact_id.
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Contact not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a contact
      tags:
      - contacts
    get:
      description: Returns a single Contact.
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.Contact'
        "404":
          description: Contact not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a contact
      tags:
      - contacts
    put:
      consumes:
      - application/json
      description: Updates a Contact. role_title must match an active configured job
        title.
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contact fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.contactForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.Contact'
        "400":
          description: Invalid body or invalid role_title
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a contact
      tags:
      - contacts
  /contacts/{id}/restore:
    post:
      description: Un-deletes a soft-deleted Contact.
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.Contact'
        "403":
          description: Not Admin/Sales Manager
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deleted contact not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Restore a deleted contact (Admin/Sales Manager only)
      tags:
      - contacts
...
  /contacts/trash:
    get:
      description: Returns soft-deleted Contacts.
      produces:
      - application/json
      responses:
        "200":
          description: Paginated contact list (data, page, per_page, total)
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not Admin/Sales Manager
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List deleted contacts (Admin/Sales Manager only)
      tags:
      - contacts
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT, e.g. "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"